auth/requirements.txt
nameko
nameko-sqlalchemy
alembic
marshmallow
psycopg2